DateRangePicker: split tabs into impl files, move the local model to desktop, tidy
- `DateRangePicker.ts` keeps the component, trigger, step buttons, popover shell and footer; the four tabs and their shared helpers move to `desktop/cmp/daterange/impl/`.
- `DateRangePickerLocalModel` moves from `cmp/daterange/impl` to `desktop/cmp/daterange/impl` - it is popover view state (tab rail, two-month calendar, footer prose), not something a mobile picker would share.
- Utils: the three pinned-period kinds resolve and step through one `periodStart()` path; the `equalDurationPrior` wrapper is gone.
- Period tab buttons share `selectableProps()`; presets file leads with its exports, helpers hoisted below.
- Comments trimmed where the code already said it; stale "month and year" docs now cover quarters.
