+++ desktop/cmp/daterange/DateRangePicker.ts
@@ -4,16 +4,11 @@
  *
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
-import {
-    DATE_RANGE_UNITS,
-    DateRangePickerModel,
-    getDateRangePresetName,
-    getDateRangeUnitLabel
-} from '@xh/hoist/cmp/daterange';
+import {DateRangePickerModel} from '@xh/hoist/cmp/daterange';
 import {
     DateRangePickerLocalModel,
     type DateRangePickerTabSpec
-} from '@xh/hoist/cmp/daterange/impl/DateRangePickerLocalModel';
+} from './impl/DateRangePickerLocalModel';
 import {box, div, filler, fragment, hbox, span, vbox} from '@xh/hoist/cmp/layout';
 import {tabContainer} from '@xh/hoist/cmp/tab';
 import {
@@ -26,17 +21,18 @@ import {
     uses
 } from '@xh/hoist/core';
 import {button, type ButtonProps} from '@xh/hoist/desktop/cmp/button';
-import {dateInput, numberInput, segmentedControl} from '@xh/hoist/desktop/cmp/input';
 import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
-import {controlGroup, popover} from '@xh/hoist/kit/blueprint';
+import {popover} from '@xh/hoist/kit/blueprint';
 import type {PopoverPosition} from '@blueprintjs/core';
-import type {LocalDate} from '@xh/hoist/utils/datetime';
-import {elemWithin, getTestId, TEST_ID} from '@xh/hoist/utils/js';
+import {elemWithin, getTestId} from '@xh/hoist/utils/js';
 import {composeRefs, splitLayoutProps, useOnResize} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
-import {isString} from 'lodash';
 import type {KeyboardEvent, ReactNode} from 'react';
+import {customTab} from './impl/CustomTab';
+import {periodTab} from './impl/PeriodTab';
+import {presetsTab} from './impl/PresetsTab';
+import {relativeTab} from './impl/RelativeTab';
 import './DateRangePicker.scss';
 
 export interface DateRangePickerProps
@@ -92,13 +88,13 @@ export interface DateRangePickerProps
 
 /**
  * A dropdown control for selecting a period of time - one compact trigger that can express
- * presets (e.g. MTD, Last 30 Days), relative lookbacks, calendar months and years, and custom
- * ranges of dates.
+ * presets (e.g. MTD, Prev 30 Days), relative lookbacks, calendar months, quarters and years, and
+ * custom ranges of dates.
  *
  * The trigger shows the applied period's label with its resolved dates, and opens a popover with a
  * tab for each of those selection shapes. The backing {@link DateRangePickerModel}'s `tabs` config
- * selects which tabs show. Preset and month/year picks commit on click. Relative and custom picks
- * are drafts until Apply, held in a local model that never touches the applied value.
+ * selects which tabs show. Preset and period picks commit on click. Relative and custom picks are
+ * drafts until Apply, held in a local model that never touches the applied value.
  *
  * App code constructs and persists the {@link DateRangePickerModel}, which owns the applied value
  * and the ranges and filters it resolves to - see that class for the full API.
@@ -284,430 +280,6 @@ const popoverContent = hoistCmp.factory<DateRangePickerLocalModel>(
     }
 );
 
-const sectionHeader = (text: string) =>
-    div({className: 'xh-date-range-picker-popover__section-hdr', item: text});
-
-//------------------
-// Tab - Presets
-//------------------
-const presetsTab = hoistCmp.factory<DateRangePickerLocalModel>(({model, testId}) => {
-    const {parentModel, singleTab} = model,
-        {value, context} = parentModel;
-
-    return vbox({
-        className: 'xh-date-range-picker-popover__tab',
-        items: [
-            sectionHeader('Presets'),
-            ...parentModel.presets.map(preset => {
-                const {token} = preset,
-                    selected = value.kind === 'preset' && value.token === token,
-                    range = parentModel.resolve({kind: 'preset', token}).current;
-                return div({
-                    key: token,
-                    className: classNames(
-                        'xh-date-range-picker-popover__preset-row',
-                        selected && 'xh-date-range-picker-popover__preset-row--selected'
-                    ),
-                    ...dataAttrs({[TEST_ID]: getTestId(testId, `preset-${token}`)}),
-                    ...clickable(() => model.commit({kind: 'preset', token})),
-                    items: [
-                        Icon.check({
-                            className: classNames(
-                                'xh-date-range-picker-popover__row-check',
-                                !selected && 'xh-date-range-picker-popover__row-check--hidden'
-                            )
-                        }),
-                        span({
-                            className: 'xh-date-range-picker-popover__row-name',
-                            item: getDateRangePresetName(preset, context)
-                        }),
-                        span({
-                            omit: singleTab,
-                            className: 'xh-date-range-picker-popover__row-range',
-                            item: parentModel.fmtRange(range)
-                        })
-                    ]
-                });
-            })
-        ]
-    });
-});
-
-//------------------
-// Tab - Relative
-//------------------
-const relativeTab = hoistCmp.factory<DateRangePickerLocalModel>(({model, testId}) => {
-    const {parentModel, relativeRange} = model;
-
-    return vbox({
-        className: 'xh-date-range-picker-popover__tab',
-        items: [
-            sectionHeader('Relative Lookback'),
-            hbox({
-                className: 'xh-date-range-picker-popover__rel-row',
-                items: [
-                    span({className: 'xh-date-range-picker-popover__rel-last', item: 'Last'}),
-                    controlGroup({
-                        items: [
-                            button({
-                                text: '−',
-                                minWidth: 20,
-                                testId: getTestId(testId, 'count-down'),
-                                onClick: () => model.stepCount(-1)
-                            }),
-                            numberInput({
-                                model,
-                                bind: 'relativeCount',
-                                min: 1,
-                                max: 999,
-                                width: 56,
-                                textAlign: 'center',
-                                commitOnChange: true,
-                                testId: getTestId(testId, 'count')
-                            }),
-                            button({
-                                text: '+',
-                                minWidth: 20,
-                                testId: getTestId(testId, 'count-up'),
-                                onClick: () => model.stepCount(1)
-                            })
-                        ]
-                    }),
-                    segmentedControl({
-                        model,
-                        bind: 'relativeUnit',
-                        testId: getTestId(testId, 'unit'),
-                        options: DATE_RANGE_UNITS.map(unit => ({
-                            value: unit,
-                            label: getDateRangeUnitLabel(unit)
-                        }))
-                    })
-                ]
-            }),
-            div({
-                className: 'xh-date-range-picker-popover__snap-box',
-                items: [
-                    // Two ways of counting, named as peers - a checkbox would frame one as a
-                    // modifier on the other, inviting the read that it widens the range. Days have
-                    // nothing to round out, so offer no choice at that grain.
-                    segmentedControl({
-                        omit: !model.snapApplies,
-                        model,
-                        bind: 'relativeSnap',
-                        fill: false,
-                        testId: getTestId(testId, 'snap'),
-                        options: [
-                            {value: false, label: 'Rolling'},
-                            {value: true, label: 'Calendar'}
-                        ]
-                    }),
-                    div({
-                        className: 'xh-date-range-picker-popover__snap-help',
-                        item: model.relativeModeHelpText
-                    })
-                ]
-            }),
-            div({
-                className: 'xh-date-range-picker-popover__preview',
-                items: [
-                    div({
-                        className: 'xh-date-range-picker-popover__preview-label',
-                        item: 'Resolves To'
-                    }),
-                    div({
-                        className: 'xh-date-range-picker-popover__preview-range',
-                        item: parentModel.fmtRange(relativeRange)
-                    }),
-                    div({
-                        className: 'xh-date-range-picker-popover__preview-days',
-                        item: model.relativePreviewCountText
-                    })
-                ]
-            })
-        ]
-    });
-});
-
-//------------------
-// Tab - Period (months, quarters, years)
-//------------------
-const MONTH_LABELS = [
-    'Jan',
-    'Feb',
-    'Mar',
-    'Apr',
-    'May',
-    'Jun',
-    'Jul',
-    'Aug',
-    'Sep',
-    'Oct',
-    'Nov',
-    'Dec'
-];
-
-const periodTab = hoistCmp.factory<DateRangePickerLocalModel>(({model, testId}) => {
-    const {parentModel, gridYear} = model,
-        {value} = parentModel,
-        yearRange = parentModel.resolve({kind: 'year', year: gridYear}).current,
-        yearSelected = value.kind === 'year' && value.year === gridYear;
-
-    return vbox({
-        className: 'xh-date-range-picker-popover__tab',
-        items: [
-            hbox({
-                className: 'xh-date-range-picker-popover__month-hdr',
-                items: [
-                    sectionHeader('Months & Years'),
-                    filler(),
-                    button({
-                        className: 'xh-date-range-picker-popover__nav-btn',
-                        icon: Icon.angleLeft(),
-                        minimal: true,
-                        disabled: !model.canYearBack,
-                        testId: getTestId(testId, 'year-back'),
-                        onClick: () => model.stepYear(-1)
-                    }),
-                    span({
-                        className: 'xh-date-range-picker-popover__year-val',
-                        item: String(gridYear)
-                    }),
-                    button({
-                        className: 'xh-date-range-picker-popover__nav-btn',
-                        icon: Icon.angleRight(),
-                        minimal: true,
-                        disabled: !model.canYearForward,
-                        testId: getTestId(testId, 'year-forward'),
-                        onClick: () => model.stepYear(1)
-                    })
-                ]
-            }),
-            div({
-                // One row per quarter: its three months, then the quarter itself - so the quarter
-                // labels the months it contains, and both are a click away.
-                className: 'xh-date-range-picker-popover__month-grid',
-                items: [1, 2, 3, 4].flatMap(quarter => {
-                    const quarterActive =
-                            value.kind === 'quarter' &&
-                            value.quarter === quarter &&
-                            value.year === gridYear,
-                        months = [quarter * 3 - 2, quarter * 3 - 1, quarter * 3];
-                    return [
-                        ...months.map(month => {
-                            const active =
-                                value.kind === 'month' &&
-                                value.month === month &&
-                                value.year === gridYear;
-                            return button({
-                                key: month,
-                                className: 'xh-date-range-picker-popover__month-btn',
-                                text: MONTH_LABELS[month - 1],
-                                // Selected reads as a solid accent fill, like the unit selector's
-                                // selected segment. Unselected keeps the outlined treatment used
-                                // elsewhere in the picker - note Hoist buttons are minimal by
-                                // default, so a solid fill needs `minimal: false`, not just
-                                // `outlined: false`.
-                                outlined: !active,
-                                minimal: !active,
-                                active,
-                                intent: active ? (model.intent ?? 'primary') : undefined,
-                                disabled: model.isMonthDisabled(month),
-                                testId: getTestId(testId, `month-${month}`),
-                                onClick: () => model.commit({kind: 'month', year: gridYear, month})
-                            });
-                        }),
-                        button({
-                            key: `q${quarter}`,
-                            className: 'xh-date-range-picker-popover__quarter-btn',
-                            text: `Q${quarter}`,
-                            outlined: !quarterActive,
-                            minimal: !quarterActive,
-                            active: quarterActive,
-                            intent: quarterActive ? (model.intent ?? 'primary') : undefined,
-                            disabled: model.isQuarterDisabled(quarter),
-                            testId: getTestId(testId, `quarter-${quarter}`),
-                            onClick: () => model.commit({kind: 'quarter', year: gridYear, quarter})
-                        })
-                    ];
-                })
-            }),
-            button({
-                className: 'xh-date-range-picker-popover__year-row',
-                outlined: !yearSelected,
-                minimal: !yearSelected,
-                active: yearSelected,
-                intent: yearSelected ? (model.intent ?? 'primary') : undefined,
-                disabled: !model.yearSelectable,
-                testId: getTestId(testId, 'year-row'),
-                onClick: () => model.commit({kind: 'year', year: gridYear}),
-                items: [
-                    Icon.calendar(),
-                    span({
-                        className: 'xh-date-range-picker-popover__year-row-label',
-                        item: String(gridYear)
-                    }),
-                    span({
-                        // As on the Presets tab - the single-tab popover is too narrow for dates.
-                        omit: model.singleTab,
-                        className: 'xh-date-range-picker-popover__row-range',
-                        item: parentModel.fmtRange(yearRange)
-                    })
-                ]
-            })
-        ]
-    });
-});
-
-//------------------
-// Tab - Custom
-//------------------
-const customTab = hoistCmp.factory<DateRangePickerLocalModel>(({model, testId}) => {
-    const {parentModel, minDate, maxDate, leftMonth, rightMonth} = model;
-
-    const field = (edge: 'start' | 'end') =>
-        vbox({
-            className: classNames(
-                'xh-date-range-picker-popover__field',
-                model.nextEdge === edge && 'xh-date-range-picker-popover__field--armed'
-            ),
-            items: [
-                div({
-                    className: 'xh-date-range-picker-popover__field-label',
-                    item: edge === 'start' ? 'Start' : 'End'
-                }),
-                dateInput({
-                    value: edge === 'start' ? model.customStart : model.customEnd,
-                    onCommit: v =>
-                        edge === 'start' ? model.commitStartInput(v) : model.commitEndInput(v),
-                    valueType: 'localDate',
-                    // Inputs need a parseable string format - a function format cannot be typed into.
-                    formatString: isString(parentModel.dateFormat)
-                        ? parentModel.dateFormat
-                        : 'YYYY-MM-DD',
-                    enablePicker: false,
-                    rightElement: null,
-                    width: '100%',
-                    minDate,
-                    maxDate,
-                    testId: getTestId(testId, `${edge}-input`)
-                })
-            ]
-        });
-
-    return vbox({
-        className: 'xh-date-range-picker-popover__tab',
-        items: [
-            hbox({
-                className: 'xh-date-range-picker-popover__custom-fields',
-                items: [
-                    field('start'),
-                    Icon.arrowRight({className: 'xh-date-range-picker-popover__field-arrow'}),
-                    field('end')
-                ]
-            }),
-            hbox({
-                className: 'xh-date-range-picker-popover__cal-nav',
-                items: [
-                    button({
-                        className: 'xh-date-range-picker-popover__nav-btn',
-                        icon: Icon.angleLeft(),
-                        minimal: true,
-                        disabled: !model.canCalBack,
-                        testId: getTestId(testId, 'cal-back'),
-                        onClick: () => model.stepCalendar(-1)
-                    }),
-                    span({
-                        className: 'xh-date-range-picker-popover__cal-title',
-                        item: leftMonth.format('MMMM YYYY')
-                    }),
-                    span({
-                        className: 'xh-date-range-picker-popover__cal-title',
-                        item: rightMonth.format('MMMM YYYY')
-                    }),
-                    button({
-                        className: 'xh-date-range-picker-popover__nav-btn',
-                        icon: Icon.angleRight(),
-                        minimal: true,
-                        disabled: !model.canCalForward,
-                        testId: getTestId(testId, 'cal-forward'),
-                        onClick: () => model.stepCalendar(1)
-                    })
-                ]
-            }),
-            hbox({
-                className: 'xh-date-range-picker-popover__cals',
-                items: [
-                    calendarMonth({model, monthStart: leftMonth}),
-                    calendarMonth({model, monthStart: rightMonth})
-                ]
-            })
-        ]
-    });
-});
-
-const DOW_LABELS = ['S', 'M', 'T', 'W', 'T', 'F', 'S'];
-
-interface CalendarMonthProps extends HoistProps<DateRangePickerLocalModel> {
-    monthStart: LocalDate;
-}
-
-const calendarMonth = hoistCmp.factory<CalendarMonthProps>(({model, monthStart}) => {
-    const {customStart, customEnd} = model;
-
-    return div({
-        className: 'xh-date-range-picker-popover__cal',
-        items: [
-            div({
-                className: 'xh-date-range-picker-popover__cal-dow',
-                items: DOW_LABELS.map((it, idx) => span({key: idx, item: it}))
-            }),
-            ...buildWeeks(monthStart).map((week, wIdx) =>
-                div({
-                    key: wIdx,
-                    className: 'xh-date-range-picker-popover__cal-week',
-                    items: week.map((day, dIdx) => {
-                        if (!day) {
-                            return div({
-                                key: dIdx,
-                                className: 'xh-date-range-picker-popover__cal-cell'
-                            });
-                        }
-                        const disabled = model.isDayDisabled(day),
-                            edge = day === customStart || day === customEnd,
-                            inRange = day > customStart && day < customEnd;
-                        return div({
-                            key: dIdx,
-                            className: classNames(
-                                'xh-date-range-picker-popover__cal-cell',
-                                edge && 'xh-date-range-picker-popover__cal-cell--edge',
-                                inRange && 'xh-date-range-picker-popover__cal-cell--in-range',
-                                disabled && 'xh-date-range-picker-popover__cal-cell--disabled'
-                            ),
-                            ...dataAttrs({'data-date': day.isoString}),
-                            item: day.format('D'),
-                            ...clickable(disabled ? null : () => model.pickDay(day))
-                        });
-                    })
-                })
-            )
-        ]
-    });
-});
-
-function buildWeeks(monthStart: LocalDate): (LocalDate | null)[][] {
-    const startOffset = monthStart.moment.day(),
-        daysInMonth = monthStart.endOfMonth().diff(monthStart, 'days') + 1,
-        cells: (LocalDate | null)[] = [];
-
-    for (let i = 0; i < startOffset; i++) cells.push(null);
-    for (let d = 0; d < daysInMonth; d++) cells.push(monthStart.add(d, 'days'));
-    while (cells.length % 7 !== 0) cells.push(null);
-
-    const weeks = [];
-    for (let i = 0; i < cells.length; i += 7) weeks.push(cells.slice(i, i + 7));
-    return weeks;
-}
-
 //------------------
 // Footer
 //------------------
@@ -757,8 +329,6 @@ const footer = hoistCmp.factory<DateRangePickerLocalModel>(
                     omit: !showApplyControls,
                     text: 'Apply',
                     outlined: true,
-                    // Follows the component's `intent`, so the footer stays coherent with the
-                    // popover's selection accent. Primary otherwise, per the Hoist standard.
                     intent: intent ?? 'primary',
                     testId: getTestId(testId, 'apply'),
                     onClick: () => model.apply()
@@ -787,28 +357,3 @@ const TAB_SPECS: DateRangePickerTabSpec[] = [
     {id: 'period', title: 'Months & Years', icon: Icon.calendarDays(), content: periodTab},
     {id: 'custom', title: 'Custom Range', icon: Icon.calendarRange(), content: customTab}
 ];
-
-//------------------
-// Helpers
-//------------------
-/** Props making a styled div act as a keyboard-accessible button. Pass null to disable. */
-const clickable = (onClick: () => void) =>
-    onClick
-        ? {
-              role: 'button',
-              tabIndex: 0,
-              onClick,
-              onKeyDown: (e: KeyboardEvent) => {
-                  if (e.key === 'Enter' || e.key === ' ') {
-                      e.preventDefault();
-                      e.stopPropagation();
-                      onClick();
-                  }
-              }
-          }
-        : {role: 'button', 'aria-disabled': true};
-
-/**
- * Data attributes for a plain element spec, typed loosely - React's HTML attribute types omit them.
- */
-const dataAttrs = (attrs: Record<string, string>) => attrs as object;
